move handleRepoResources to its own file
